Move data related files to tests/data dir
Also, introduce a base class for implementing tests.
